rename: markdownToJsx.js to markup.js
