+++ src/app/categories/[tag]/page.jsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 
 import { DOCS, EXTENSION } from '@/constants/path';
 import { readTagTree } from '@/utils/fs/tagTree';
-import { markdownToText } from '@/utils/markdownToJsx';
+import { markdownToText } from '@/utils/markup';
 
 const { mdRegExp } = EXTENSION;
